requirements.txt
pandas>=2.0.0
openpyxl>=3.0.0
xlrd>=2.0.0
pyparsing>=3.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
python-dateutil>=2.8.0

# UI Dependencies
customtkinter>=5.2.0
Pillow>=10.0.0
matplotlib>=3.7.0
